Add support for linkable modules
This change adds support for writing wasm modules such
that they are compatible with the proposed linking spec:
https://github.com/WebAssembly/tool-conventions/blob/master/Linking.md

Basically this means that wast2wasm will generate and extra
'reloc' section and will pad any LEBs in the code section
that might require relocation.

This mode is activated by passing the -c flag to wast2wasm.

Also, start work on wasm-link which should be able to link
such modules together.